Reduce command now works.
Needed to fix an issue with theano functions. Also needed to fix
the way we were loading in data, the conversion to float was not
actually happening. Furthermore, there was an issue with
the normalization/standardization: it would change the datatype
from float32 to float64. Added a hack to fix this.