Preparing pipeline for new GO architectures.
This required:
- Changing the way grouping matrices (now called adjacency matrices) are
created. Now, rows=input nodes and columns=output nodes.
- Also required ensuring that the order of the groups in a grouping file
is preserved when converted to an adjacency matrix
- Also needed to be ably to handle groups with no genes (empty groups
should still have nodes, but should not be connected to any input genes).
This is important for GO, because while the disconnected node will not
contribute to the function learned, its presence is required to keep the expected
length and order of the layer, which following layers may depend on.