Can now run reduction without metadata
Before, you had to also supply labels and accessions for each sample,
and gene symbols for each gene. Now, you only have to provide those if
you would like to also save them in the reduced version of the data.

Do this using the --save_meta flag in the reduce command.