Data_prep is more conserving with the data
Data_prep is more of a proper command line tool, with usage help.

A new option '--uniq_lvl' allows the user to specify the level of
uniqueness they require in the cell ontology term mappings for a
sample. Before, the script mandated that only cells that map to a
single term are kept, and trashed other "non-unique" cells.

Now, the user can specify a minimum Jaccard distance that allows them
to tradeoff the number of cells with the uniqueness of their labeling.