Rolling back some features of last commit on data_prep
Choosing which of two terms to pick that overlap too much leads to
the problem of finding a bipartite subgraph, and the maximal bipartite
subgraph problem is NP-Complete, and approximation is challenging.

Solving this problem is also not productive for me, since picking
cell ontology terms still needs the human touch.

For now, we'll just print out the Jaccard distance analysis,
and manually make decisions about which nodes to remove.