PCA+LogisticRegression for comparison + more
Should compare the status quo (PCA) to NNs via the performance
of a multinomial Logistic Regression classifier for cell type
identification. This is because the last hidden layer of neural net
and the PCs can both be viewed as embeddings, and then the output
layer of the NN is basically multinomial LR (softmax activation).

Also made NN training use a fit_generator to iterate over batches
of data, and give flexibility over when to end an epoch, to allow
for more frequent updates, important as our dataset is growing.